+++ Stitch/Graph/Node/Layer/Type/Media/3DModelLayerNode/Nodes/ConeLayerNode.swift
