+++ Stitch/Graph/Node/Layer/Type/Media/3DModelLayerNode/Nodes/CylinderLayerNode.swift
