+++ Stitch/Graph/Node/Layer/Type/Media/3DModelLayerNode/Nodes/Model3DLayerNode.swift
