+++ Stitch/Graph/Node/Port/Model/PortValue/Type/Media/ObjectTypes/AR/StitchEntity.swift
@@ -133,29 +133,6 @@ extension StitchEntity {
         self.transform = newMatrix
         self.containerEntity._applyMatrix(newMatrix: newMatrix)
     }
-
-//   @MainActor func createSCNScene(layerViewModel: LayerViewModel) throws -> SCNScene {
-//       switch self.type {
-//       case .importedMedia(let url):
-//           return try SCNScene(url: url)
-//       default:
-//           fatalErrorIfDebug()
-//           return scene
-//       }
-//   }
-//    
-//    @MainActor private func buildSCNScene(from scene: SCNScene,
-//                                          layerViewModel: LayerViewModel) {
-//        guard let geometry = self.type.createSCNGeometry(layerViewModel: layerViewModel) else {
-//            return
-//        }
-//        
-//        let node = SCNNode(geometry: geometry)
-//        scene.rootNode.addChildNode(node)
-//        
-//        self.type.updateSCNScene(from: scene,
-//                                 layerViewModel: layerViewModel)
-//    }
     
     @MainActor
     func update(layerViewModel: LayerViewModel) {
